Restructure the memory layout of each process
The process memory layout can be illustrated below

Non Accessable First Page | 1024 words
Stack | 1024 words
text
data
bss
Initial Heap | 2048 words

+++ lib/ansi/env.c
@@ -5,6 +5,6 @@
 const char** _environ = NULL;
 
 void init_environ(){
-    _environ = *((const char ***)ALIGN1K(get_sp()));
+    _environ = *((const char ***)ALIGN1K((int)get_sp()));
     // printf("env %x \n", (unsigned int)_environ + 0x5800);
 }
